Adds robots.txt and sitemap.xml
Adds robots.txt and sitemap.xml files to improve SEO and guide search engine crawlers.

The robots.txt file defines which parts of the site should not be accessed by search engine crawlers.
The sitemap.xml file provides a list of all the pages on the site, helping search engines to index the content more efficiently.